bug fix in DeleteTaskCommandHandler

+++ src/API/Todo.Api/Todo.Api.Application/Features/Tasks/Commands/DeleteTask/DeleteTaskListCommandHandler.cs
@@ -1,6 +1,7 @@
 ﻿using AutoMapper;
 using MediatR;
 using Todo.Api.Application.Contracts.Persistence;
+using Todo.Api.Application.Exceptions;
 using Entities = Todo.Api.Domain.Entities;
 
 namespace Todo.Api.Application.Features.Task.Commands.DeleteTask
@@ -17,6 +18,11 @@ namespace Todo.Api.Application.Features.Task.Commands.DeleteTask
         public async Task<Unit> Handle(DeleteTaskCommand request, CancellationToken cancellationToken)
         {
             var taskListToDelete = await _taskListRepository.GetByIdAsync(request.Id);
+
+            if (taskListToDelete == null)
+            {
+                throw new NotFoundException(nameof(Entities.Task), request.Id);
+            }
 
             await _taskListRepository.DeleteAsync(taskListToDelete);
